rename: 역할별로 패키지 구조 및 이름 변경
