+++ src/main/java/baseball/io/UserIo.java
@@ -1,4 +1,4 @@
-package baseball.user;
+package baseball.io;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.Arrays;
